ThreadXY
- Coordinateshift for sensor staat nu in een function "shiftCoordinate(return)"
- Coordinatenstelsel naar doubles omgezet en stock positie gewijzigd
- moveXY opgesplitst in 2 threads

+++ Tic Tac Toe/src/OpdrachtZet.java
@@ -5,7 +5,7 @@ public class OpdrachtZet extends Opdracht {
 	//mogelijkheid 1: met geen methode opdracht in de klasse opdracht
 	
 	//constructor, hoe definieren we current? is eindepositie vorige zet
-	OpdrachtZet(int [] s, int [] e){
+	OpdrachtZet(double [] s, double [] e){
 		start=s;
 		end=e;
 	}
